requirements.txt
torch==2.7.1
transformers==4.57.5
accelerate==1.10.1
vllm==0.10.1.1
safetensors==0.6.2
tokenizers==0.22.0
sentencepiece==0.2.1
lm_eval==0.4.9.1
datasets==3.6.0
numpy==1.26.4
pandas==2.3.0
scipy==1.16.0
scikit-learn==1.6.1
statsmodels==0.14.5
matplotlib==3.10.0
seaborn==0.13.2
nltk==3.9.1
regex==2024.11.6
rouge_score==0.1.2
sacrebleu==2.5.1
sentence-transformers==5.2.0
huggingface-hub==0.34.4
tqdm==4.67.1
PyYAML==6.0.2
filelock==3.18.0
jsonlines==4.0.0
dill==0.3.8